# lib/api-zod/src/generated/types/agentWorkCreateInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Personal Secretary API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentWorkCreateInputCondition } from './agentWorkCreateInputCondition';
import type { AgentWorkCreateInputKind } from './agentWorkCreateInputKind';
import type { AgentWorkCreateInputSchedule } from './agentWorkCreateInputSchedule';
import type { AgentWorkCreateInputSource } from './agentWorkCreateInputSource';

export interface AgentWorkCreateInput {
  kind: AgentWorkCreateInputKind;
  /**
     * @minLength 1
     * @maxLength 200
     */
  title: string;
  /**
     * @maxLength 2000
     * @nullable
     */
  description?: string | null;
  source?: AgentWorkCreateInputSource;
  condition?: AgentWorkCreateInputCondition;
  schedule?: AgentWorkCreateInputSchedule;
  /** @nullable */
  nextRunAt?: Date | null;
}