# lib/api-zod/src/generated/types/agentWorkDetails.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Personal Secretary API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentWork } from './agentWork';
import type { AgentWorkEvent } from './agentWorkEvent';
import type { AgentWorkEvidence } from './agentWorkEvidence';
import type { AgentWorkRun } from './agentWorkRun';

export interface AgentWorkDetails {
  work: AgentWork;
  runs: AgentWorkRun[];
  events: AgentWorkEvent[];
  evidence: AgentWorkEvidence[];
}