# lib/api-zod/src/generated/types/agentWorkEvent.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Personal Secretary API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentWorkEventMetadata } from './agentWorkEventMetadata';

export interface AgentWorkEvent {
  id: string;
  workId: string;
  /** @nullable */
  runId: string | null;
  eventType: string;
  actorType: string;
  /** @nullable */
  actorId: string | null;
  summary: string;
  metadata: AgentWorkEventMetadata;
  occurredAt: Date;
  createdAt: Date;
  [key: string]: unknown;
 }