# lib/api-zod/src/generated/types/agentWorkEvidence.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Personal Secretary API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentWorkEvidenceRetentionClass } from './agentWorkEvidenceRetentionClass';
import type { AgentWorkEvidenceSnapshot } from './agentWorkEvidenceSnapshot';

export interface AgentWorkEvidence {
  id: string;
  workId: string;
  runId: string;
  snapshotHash: string;
  snapshot: AgentWorkEvidenceSnapshot;
  retentionClass: AgentWorkEvidenceRetentionClass;
  /** @nullable */
  expiresAt: Date | null;
  createdAt: Date;
}