# lib/api-zod/src/generated/types/agentWorkRun.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Personal Secretary API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentWorkRunStatus } from './agentWorkRunStatus';
import type { AgentWorkRunVerification } from './agentWorkRunVerification';

export interface AgentWorkRun {
  id: string;
  workId: string;
  attempt: number;
  status: AgentWorkRunStatus;
  idempotencyKey: string;
  /** @nullable */
  leaseExpiresAt: Date | null;
  /** @nullable */
  startedAt: Date | null;
  /** @nullable */
  completedAt: Date | null;
  /** @nullable */
  verification: AgentWorkRunVerification;
  /** @nullable */
  error: string | null;
  createdAt: Date;
  updatedAt: Date;
}