# lib/api-zod/src/generated/types/agentWorkStatusInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Personal Secretary API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentWorkStatusInputFrom } from './agentWorkStatusInputFrom';
import type { AgentWorkStatusInputTo } from './agentWorkStatusInputTo';

export interface AgentWorkStatusInput {
  from: AgentWorkStatusInputFrom;
  to: AgentWorkStatusInputTo;
  /** @maxLength 500 */
  reason?: string;
}